requirements.txt
Pillow==4.3.0
Keras==2.0.8
tensorflow==1.3.0
opencv-python==3.3.0.10
pandas==0.20.3
requests==2.18.4